feat(mobile): Library is a reader-first front door — resume, then everything else
A logged-in user opening Library met "Open filters", their own email
address and a Saved/Uploads tab bar before they met the book they were
reading. The only personal content above the fold was a shelf whose
payload (/me/library/shelves) carries no chapterSlug, so tapping it could
only open a detail page — never resume.

Now the screen is a strict priority stack, not a feed:
  1. ResumeHero      — one full-width card, one tap, straight into the
                       reader at the saved chapter
  2. JumpBackInRail  — the rest of the in-progress books, also resuming
  3. Vocabulary
  4. Shelves, then the browsing chrome, then the grid

rankContinueReading() in @textstack/shared is the enabling piece: the LWW
server/local merge already existed for the single "Continue Reading" pick
and simply threw away everything but the winner. pickContinueReadingBook
is now rank(...)[0], so the hero and the rail can never disagree. No
backend change; all 23 existing merge tests pass unmodified.

FirstBookState replaces the generic "your library is empty / browse the
catalog" for accounts with no books. It has to live at screen level: the
list's own guard suppresses the shelves header in exactly the all-empty
case, so an empty state nested there would never render.

Leaving the reader now invalidates the shelves TTL cache — that is the one
event which always makes the front door stale.

app/(tabs)/library.tsx drops 970 → 245 lines; SavedList, UploadsList and
the shared stylesheet move out verbatim first, so this diff is readable.

Home still shows its own ContinueReadingCard. That duplication is
deliberate and temporary: the Library front door ships first and proves
itself before the tab bar changes.
